moved SqlFragment and Proxy from Types
